style: sections modified and page address changed to /pearl
